CompoundErrors: Backport rename of error collector
TaskValidationErrors is part of the public API for tasks and needs to be
renamed before tasks comes out of beta.
